eyedisease_detector/requirements.txt
Django
tensorflow
pillow
numpy